TESTS CASES COMPLETED

+++ src/main/java/com/Employes/Employedetail/service/employeeservice.java
@@ -23,31 +23,42 @@ public class employeeservice extends employeevalidate{
     DesignationRepo depo;
 
 //GET THE DETAILS OF EMPLOYEES USING ID
-        public Map<String, Object> getUserDetails(Integer eid)
-        {
+        public ResponseEntity getUserDetails(Integer eid) {
 
-            Employee manager=null;
-            List<Employee> colleagues=null;
-            Map<String,Object> map=new LinkedHashMap<>();
-
-            boolean userExists=userExists(eid);
-            Employee emp=repo.findByEid(eid);
-
-            map.put("Employee",emp);
-
-            if(emp.getManager()!=null) {
-                manager = repo.findByEid(emp.getManager());
-                map.put("Manager",manager);
-
-                colleagues=repo.findAllByManagerAndEidIsNot(emp.getManager(),emp.getEid());
-                map.put("Colleagues",colleagues);
+            Employee manager;
+            List<Employee> colleagues;
+            Map<String, Object> map = new LinkedHashMap<>();
+            boolean userExists = false;
+            if (eid != null && eid > 0) {
+                userExists = userExists(eid);
             }
+            else if (eid < 0)
+            {
+                return new ResponseEntity<>("INVALID_ID", HttpStatus.BAD_REQUEST);
+            }
+            if (userExists) {
+                Employee emp = repo.findByEid(eid);
+                map.put("id", emp.getEid());
+                map.put("name", emp.getEmpname());
+                map.put("jobTitle", emp.getJobtitle());
+                map.put("Employee", emp);
 
-            List<Employee> reporting=repo.findAllByManagerAndEidIsNot(emp.getEid(),emp.getEid());
-            if(reporting.size()!=0)
-                map.put("Reporting To",reporting);
+                if (emp.getManager() != null) {
+                    manager = repo.findByEid(emp.getManager());
+                    map.put("manager", manager);
+                    colleagues = repo.findAllByManagerAndEidIsNotOrderByDesignation_levelAscEmpnameAsc(emp.getManager(), emp.getEid());
+                    map.put("colleagues", colleagues);
+                }
 
-            return map;
+                List<Employee> reporting = repo.findAllByManagerAndEidIsNotOrderByDesignation_levelAscEmpnameAsc(emp.getEid(), emp.getEid());
+                if (reporting.size() != 0)
+                    map.put("subordinates", reporting);
+
+                return new ResponseEntity<>(map, HttpStatus.OK);
+            }
+            else {
+                return new ResponseEntity<>("EMP_NOT_EXISTS", HttpStatus.NOT_FOUND);
+            }
         }
 //GET DETAILS OF ALL EMPLOYEES
         public ResponseEntity getAll()
@@ -56,16 +67,19 @@ public class employeeservice extends employeevalidate{
             if(hasData(list))
                 return new ResponseEntity<>(list, HttpStatus.OK);
             else
-                return new ResponseEntity<>("No Records Found",HttpStatus.NOT_FOUND);
+                return new ResponseEntity<>("No Records Found",HttpStatus.BAD_REQUEST);
         }
 //DELETE THE CURRENT EMPLOYEE
         public ResponseEntity deleteUser(Integer eid)
         {
+            if(eid<0){
+                return new ResponseEntity("INVALID_ID",HttpStatus.BAD_REQUEST);
+            }
             boolean userExists=userExists(eid);
             if(userExists)
             {
                 Employee emp=repo.findByEid(eid);
-                if(emp.getDesignation().getDesignation().equals("DIRECTOR"))
+                if(emp.getDesignation().getDesignation().equals("Director"))
                 {
                     List<Employee> list=repo.findAllByManager(emp.getEid());
                     if(hasData(list))
@@ -77,7 +91,7 @@ public class employeeservice extends employeevalidate{
                     {
                         //Able to delete
                         repo.delete(emp);
-                        return new ResponseEntity("Deleted Successfully",HttpStatus.OK);
+                        return new ResponseEntity("Deleted Successfully",HttpStatus.NO_CONTENT);
                     }
                 }
                 else
@@ -90,7 +104,7 @@ public class employeeservice extends employeevalidate{
                         repo.save(employee);
                     }
                     repo.delete(emp);
-                    return new ResponseEntity("Deleted Successfully",HttpStatus.OK);
+                    return new ResponseEntity("Deleted Successfully",HttpStatus.NO_CONTENT);
                 }
             }
             else
@@ -104,6 +118,11 @@ public class employeeservice extends employeevalidate{
             String empName=employee.getName();
             String desg=employee.getJobTitle();
             Integer parentId=employee.getManagerId();
+           if(parentId!=null){
+               if(parentId.intValue()<0){
+                   parentId=null;
+               }
+           }
           if(empName==null && desg==null && parentId==null){
               return new ResponseEntity("Please Enter valid data",HttpStatus.BAD_REQUEST);
           }
@@ -123,16 +142,16 @@ public class employeeservice extends employeevalidate{
          if(parentId==null) {
                 Employee director = repo.findByManager(null);
                 if (director != null) {
-                    return new ResponseEntity("Director Already Exists ParentId cannot be NULL", HttpStatus.FORBIDDEN);
+                    return new ResponseEntity("Director Already Exists ParentId cannot be NULL", HttpStatus.BAD_REQUEST);
                 }
                 else
                 {
-                    if(desg.equals("DIRECTOR"))
+                    if(desg.equals("Director"))
                     {
                         Designation designation=depo.findByDesignation(desg);
                         Employee emp=new Employee(designation,parentId,empName);
                         repo.save(emp);
-                        return new ResponseEntity("Employee added successfully",HttpStatus.OK);
+                        return new ResponseEntity(emp,HttpStatus.CREATED);
                     }
                     else
                     {
@@ -151,7 +170,7 @@ public class employeeservice extends employeevalidate{
                 else
                 {
                     Designation designation=depo.findByDesignation(desg);
-                    System.out.println(desg);
+
                     float currentLevel=designation.getLevel();
 
                     Employee parentRecord=repo.findByEid(parentId);
@@ -161,11 +180,11 @@ public class employeeservice extends employeevalidate{
                     {
                         Employee emp=new Employee(designation,parentId,empName);
                         repo.save(emp);
-                        return new ResponseEntity("Employee added successfully",HttpStatus.OK);
+                        return new ResponseEntity(emp,HttpStatus.CREATED);
                     }
                     else
                     {
-                        return new ResponseEntity(desg+" cannot be child of "+parentRecord.getDesignation().getDesignation(),HttpStatus.BAD_REQUEST);
+                        return new ResponseEntity("INVALID PARENT",HttpStatus.BAD_REQUEST);
                     }
                 }
             }
@@ -174,46 +193,56 @@ public class employeeservice extends employeevalidate{
         public ResponseEntity updateUser(Integer eid, putRequest emp)
         {
             //User Is Present
+            if(eid<0){
+                return new ResponseEntity("INVALID_ID",HttpStatus.BAD_REQUEST);
+            }
+
+            if(emp.getName()==null && emp.getJobTitle()==null && emp.getManagerId()==null && emp.isReplace()==false){
+                return new ResponseEntity("NO Data",HttpStatus.BAD_REQUEST);
+            }
+            if(emp.getManagerId()!=null){
+               if(!isValidId(emp.getManagerId())){
+                   return new ResponseEntity("INVALID_MANAGER",HttpStatus.BAD_REQUEST);
+                }
+            }
 
             if(userExists(eid))
             {
                 String userDesignation=emp.getJobTitle();
 
-                if(emp.isReplace())
-                {
-                    if(userDesignation==null)
-                        return new ResponseEntity("Designation cannot be Null",HttpStatus.BAD_REQUEST);
-                    else
-                    {
+                if(emp.isReplace()) {
+                    if (userDesignation == null)
+                        return new ResponseEntity("Designation cannot be Null", HttpStatus.BAD_REQUEST);
+                    else {
                         if (!isDesignationValid(userDesignation))
-                            return new ResponseEntity("Designation does't exists! Please enter valid designation",HttpStatus.BAD_REQUEST);
+                            return new ResponseEntity("Designation does't exists! Please enter valid designation", HttpStatus.BAD_REQUEST);
                     }
-                    if(!isValid(emp.getName())){
-                        return new ResponseEntity("ENTER VALID NAME",HttpStatus.BAD_REQUEST);
+                    if (!isValid(emp.getName())) {
+                        return new ResponseEntity("ENTER VALID NAME", HttpStatus.BAD_REQUEST);
                     }
-                    Integer parent=null;
-                    Employee employee=repo.findByEid(eid);
-                    if(isGreaterThanEqualCurrentDesignation(eid,userDesignation) && isSmallerThanParent(eid,userDesignation))
-                    {
-                        parent=employee.getManager();
+                    Integer parent = null;
+                    Employee employee = repo.findByEid(eid);
+                    if (isGreaterThanEqualCurrentDesignation(eid, userDesignation) && isSmallerThanParent(eid, userDesignation)) {
+                        parent = employee.getManager();
                         repo.delete(employee);
-                        Employee tempEmployee=new Employee(depo.findByDesignation(userDesignation),parent,emp.getName());
+                        Employee tempEmployee = new Employee(depo.findByDesignation(userDesignation), parent, emp.getName());
                         repo.save(tempEmployee);
-                        List<Employee> list=repo.findAllByManager(eid);
-                        for(Employee empTemp:list)
-                        {
+                        List<Employee> list = repo.findAllByManager(eid);
+                        for (Employee empTemp : list) {
                             empTemp.setManager(tempEmployee.getEid());
                             repo.save(empTemp);
                         }
-                        return new ResponseEntity("User Replaced",HttpStatus.OK);
-                    }
-                    else
-                        return new ResponseEntity(employee.getDesignation().getDesignation()+" cannot be replaced with "+userDesignation,HttpStatus.BAD_REQUEST);
+                        return getUserDetails(tempEmployee.getEid());
+                    } else
+                        return new ResponseEntity(" cannot be replaced with ", HttpStatus.BAD_REQUEST);
                 }
                 else
                 {
+
+                    userDesignation=emp.getJobTitle();
                     Employee employee=repo.findByEid(eid);
                     Integer parentId=emp.getManagerId();
+
 
                     if(userDesignation!=null)
                     {
@@ -265,7 +294,7 @@ public class employeeservice extends employeevalidate{
                         employee.setEmpname(emp.getName());
                     }
                     repo.save(employee);
-                    return new ResponseEntity("User Updated",HttpStatus.OK);
+                    return getUserDetails(eid);
                 }
             }
             //USER IS NOT PRESENT
